Don't use/verify Cinder capabilities if no backend
Cinder volume and backend should have a supported storage backend
configured in order to use and verify volumes/snapshots with a
test VM that we start before adoption.

When the backend is not Ceph (the only backend being tested in upstream
CI), do not attach volumes and do not create snapshots for the test VM.
Neither verify if those are still present post-adoption.

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>